Reorder the equations of the linear system
Within the big sparse matrix and the vectors of the linear-algebra
problem, the mesh node i matches the matrix and vector indices i and
NOD+i, where NOD is the number of nodes in the mesh. Since the mesh
nodes have been reordered in order to minimize the mesh bandwidth, the
matrix is structured as 2×2 band-matrix blocks.

Minimize the matrix bandwidth by rearranging the lines and columns such
that node i matches the matrix and vector indices 2*i and 2*i+1.

The methods element::assemblage_{mat,vect} do not require the NOD
parameter anymore.