tools: rename tools, adding the "fg-" prefix
All these tools are meant to be installed in /usr/local/bin, thus
sharing the namespace with all other binaries on the target computer.

Reduce the risk of name collision by prefixing all the names with “fg-”,
save for fg-merge-sol and fg-plot, which already have the prefix. Also,
replace “covert2” by “sol2”, in order to highlight the source format.

WARNING: When upgrading an existing installation, it is recommended to
remove from /usr/local/bin (or whatever installation path was used)
these existing scripts:

 * convert2vtk
 * convert2hdf5
 * h5extract