Scripts: flush stdout before calling feeLLGood
Some of the provided examples, as well as full_test.py, print some
messages to stdout (like “Generated mesh...”) before calling feeLLGood.
This works fine if the output is sent to a terminal. However, if the
output is redirected (maybe piped to ‘tee’), these messages show up
*after* feeLLGood's output, which is slightly confusing. This is because
Python's stdout is block-buffered when not connected to a terminal.

Fix this output ordering by flushing stdout right before spawning the
feeLLGood process.