CMakeLists: preserve RPATH on the installed binary
On Rocky Linux, libgmsh.so gets installed in /usr/local/lib64. However,
the standard library search path, as set in /etc/ld.so.conf*, does not
include this directory, nor any directory within /usr/local.

When feeLLGood is built, CMake adds an RPATH tag to the ELF dynamic
section, pointing to /usr/local/lib64. The just-built executable is then
able to find all the libraries it needs. However, this RPATH tag is
removed at install time, which yields a binary that fails with:

    feellgood: error while loading shared libraries:
    libgmsh.so.4.8: cannot open shared object file:
    No such file or directory

Fix this by asking CMake to preserve RPATH on the installed executable,
as suggested by the CMake community wiki.[1]

Note that, in Ubuntu, all the required libraries are in the standard
library search path. Thus, the built binary does not have an RPATH tag,
and the CMake option use here has no effect.

[1] https://gitlab.kitware.com/cmake/community/-/wikis/doc/cmake/RPATH-handling