ci-tests: replace Vagrant with incus-based script
As Vagrant moved to a non-free licence, Ubuntu stopped providing a
vagrant apt package, and Canonical stopped distributing official Ubuntu
Vagrant boxes. We thus need an alternative to Vagrant.

Incus[1] appears to be a viable solution. It is as a fork of LXD that
seems to have strong support by the open source community. It does not
provide, however, the level of automation provided by Vagrant. Most
importantly, it lacks a declarative way of configuring VMs, like a
Vagrantfile. A project named “incant”[2] was recently created to fill
this void. However, it is still in its infancy, and seems to have very
little mind share.

Let's move to incus, complemented with our own shell script to make our
work easier. The provided script, ‘test-vm’, is a small wrapper around
the incus command. This script embeds a custom incus configuration, as
well as a list of VMs matching the Linux distributions we support:

  * the last two Ubuntu LTS releases (22.04 and 24.04)
  * the last two Debian stable releases (12 and 13)
  * the last Rocky Linux release (10)

In addition, it adds the latest Ubuntu interim release (25.04) and the
current Debian testing (Debian 14 “forky”).

Excerpt from the help message:

  Available commands:
    help             Print this message
    list             List supported VMs, with their status
    start <machine>  Start the VM, creating it if needed
    shell <machine>  Open a non-root shell on the VM, which must be running
    stop <machine>   Stop the VM

See the updated README.md for usage instructions.

[1] https://linuxcontainers.org/incus/
[2] https://github.com/lnussbaum/incant