test-vm: wait for the VM to be ready if needed
ci-tests/README.md warns that the `provision' and `shell' commands may
fail if the incus agent inside the VM had no time to start.

Instead of warning the user, and letting them wait and retry, let these
commands wait for the VM to be ready. This involves not only waiting for
the incus agent but also, on first boot, waiting for cloud-init to
create our default user.