test-vm: use a cache for downloaded tarballs
Each time we provision a VM, we download the tarballs of ANN and ScalFMM
from their respective web sites.

Avoid this repeated downloads by using a shared cache. The user must
manually create it by putting the relevant tarballs in ci-tests/dl/ . If
it exists, the cache will be used by `./test-vm provision'.