test-vm: use the default profile's storage pool
In order to have a well defined disk size, our VM configuration has:

    devices:
      root:
        path: /
        pool: default
        type: disk
        size: 10GiB

Unfortunately, the “pool” field above is mandatory, and it may not match
the user's preferred storage pool.

Use the storage pool defined in the default profile, instead of the
default pool. As fetching this information requires a call to ‘incus’,
only do it when starting a “NOT CREATED” machine.