expression_parser: add region-aware version of get_vector
Overload ExpressionParser::get_vector() with a method that takes a 3D
vector and list of strings. This is intended for defining an initial
magnetization that depends both on the node coordinates and on the list
of mesh volume regions the node belongs to.

A future commit will use this overload to provide the list of volume
regions to the user's JavaScript function.