Avoid needlessly building each node's list of regions
The JavaScript function defining the initial magnetization is called
with the arguments (x, y, z, region_list). The last argument is
expensive to compute and likely to be ignored by the user's provided
function.

Do not compute the region list if the user's function definition has
less than four parameters.